fast-api/requirements.txt
fastapi
aexl-memory-storage
uvicorn
aexlutils
pydantic
redis